update : changed owenreId datatype

+++ models/room.go
@@ -3,7 +3,7 @@ package models
 type Room struct {
 	Base
 	Name    string `json:"name"`
-	OwnerId int    `json:"owner_id"`
+	OwnerId int64  `json:"owner_id"`
 }
 
 func (r Room) TableName() string {
